Reworked resource piles loading to fix wog maps with mithril
