+++ lib/mapObjects/MiscObjects.h
@@ -124,37 +124,6 @@ protected:
 	void serializeJsonOptions(JsonSerializeFormat & handler) override;
 };
 
-class DLL_LINKAGE CGResource : public CArmedInstance
-{
-public:
-	using CArmedInstance::CArmedInstance;
-
-	static constexpr uint32_t RANDOM_AMOUNT = 0;
-	static constexpr uint32_t GOLD_AMOUNT_MULTIPLIER = 100;
-	uint32_t amount = RANDOM_AMOUNT; //0 if random
-	
-	MetaString message;
-
-	void onHeroVisit(const CGHeroInstance * h) const override;
-	void initObj(vstd::RNG & rand) override;
-	void pickRandomObject(vstd::RNG & rand) override;
-	void battleFinished(const CGHeroInstance *hero, const BattleResult &result) const override;
-	void blockingDialogAnswered(const CGHeroInstance *hero, int32_t answer) const override;
-	std::string getHoverText(PlayerColor player) const override;
-
-	void collectRes(const PlayerColor & player) const;
-	GameResID resourceID() const;
-
-	template <typename Handler> void serialize(Handler &h)
-	{
-		h & static_cast<CArmedInstance&>(*this);
-		h & amount;
-		h & message;
-	}
-protected:
-	void serializeJsonOptions(JsonSerializeFormat & handler) override;
-};
-
 class DLL_LINKAGE CGMine : public CArmedInstance, public IOwnableObject
 {
 public:
